Update appeal unit tests
- Split some appeals code and tests to AppealsRepository
