+++ spec/models/appeal_spec.rb
@@ -1,4 +1,4 @@
-describe Appeal do
+describe Appeal, focus: true do
   context "#documents_match?" do
     let(:nod_document) { Document.new(type: "NOD", received_at: 3.days.ago) }
     let(:soc_document) { Document.new(type: "SOC", received_at: 2.days.ago) }
@@ -12,6 +12,12 @@ describe Appeal do
         documents: [nod_document, soc_document, form9_document]
       )
     end
+
+    before do
+      @old_repo = Appeal.repository
+      Appeal.repository = Fakes::AppealRepository
+    end
+    after { Appeal.repository = @old_repo }
 
     subject { appeal.documents_match? }
 
@@ -61,121 +67,7 @@ describe Appeal do
     end
   end
 
-  context ".normalize_vacols_date" do
-    subject { Appeal.normalize_vacols_date(datetime) }
-
-    context "when datetime is nil" do
-      let(:datetime) { nil }
-      it { is_expected.to be_nil }
-    end
-
-    context "when datetime is in a non-UTC timezone" do
-      before { Time.zone = "America/Chicago" }
-      let(:datetime) { Time.new(2013, 9, 5, 16, 0, 0, "-08:00") }
-      it { is_expected.to eq(Time.zone.local(2013, 9, 6)) }
-    end
-  end
-
-  context ".from_records" do
-    before { Timecop.freeze(Time.utc(2015, 1, 1, 12, 0, 0)) }
-    after { Timecop.return }
-
-    let(:case_record) do
-      OpenStruct.new(
-        bfcorlid: "VBMS-ID",
-        bfac: "4",
-        bfso: "Q",
-        bfpdnum: "INSURANCE-LOAN-NUMBER",
-        bfdrodec: 11.days.ago,
-        bfdnod: 10.days.ago,
-        bfdsoc: 9.days.ago,
-        bfd19: 8.days.ago,
-        bfssoc1: 7.days.ago,
-        bfssoc2: 6.days.ago,
-        bfha: "6",
-        bfhr: "1",
-        bfregoff: "DSUSER",
-        bfdc: "M"
-      )
-    end
-
-    let(:correspondent_record) do
-      OpenStruct.new(
-        snamef: "Phil",
-        snamemi: "J",
-        snamel: "Johnston",
-        sspare1: "Chris",
-        sspare2: "M",
-        sspare3: "Johnston",
-        susrtyp: "Brother"
-      )
-    end
-
-    let(:folder_record) do
-      OpenStruct.new(
-        tivbms: "Y"
-      )
-    end
-
-    subject do
-      Appeal.from_records(
-        case_record: case_record,
-        correspondent_record: correspondent_record,
-        folder_record: folder_record
-      )
-    end
-
-    it do
-      is_expected.to have_attributes(
-        vbms_id: "VBMS-ID",
-        type: "Reconsideration",
-        file_type: "VBMS",
-        representative: "Catholic War Veterans",
-        veteran_first_name: "Phil",
-        veteran_middle_initial: "J",
-        veteran_last_name: "Johnston",
-        appellant_first_name: "Chris",
-        appellant_middle_initial: "M",
-        appellant_last_name: "Johnston",
-        appellant_relationship: "Brother",
-        insurance_loan_number: "INSURANCE-LOAN-NUMBER",
-        notification_date: Appeal.normalize_vacols_date(11.days.ago),
-        nod_date: Appeal.normalize_vacols_date(10.days.ago),
-        soc_date: Appeal.normalize_vacols_date(9.days.ago),
-        form9_date: Appeal.normalize_vacols_date(8.days.ago),
-        ssoc_dates: [
-          Appeal.normalize_vacols_date(7.days.ago),
-          Appeal.normalize_vacols_date(6.days.ago)
-        ],
-        hearing_type: :video_hearing,
-        hearing_requested: true,
-        hearing_held: true,
-        regional_office_key: "DSUSER",
-        merged: true
-      )
-    end
-
-    context "No appellant listed" do
-      let(:correspondent_record) do
-        OpenStruct.new(
-          snamef: "Phil",
-          snamemi: "J",
-          snamel: "Johnston",
-          susrtyp: "Brother"
-        )
-      end
-
-      it { is_expected.to have_attributes(appellant_relationship: "") }
-    end
-  end
-
-  context ".find" do
-    class FakeRepo
-      def self.find(_id)
-        Appeal.new(representative: "Shane's VSO")
-      end
-    end
-
+  context ".find_or_create_by_vacols_id" do
     before do
       @old_repo = Appeal.repository
       Appeal.repository = FakeRepo
